# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.DataSourceEnvResp:
    properties:
      datasource_id:
        type: integer
      dsn:
        type: string
      env:
        type: string
      gateway_id:
        type: integer
      id:
        type: integer
    type: object
  http.DataSourceListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.DataSourceResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.DataSourceResp:
    properties:
      created_at:
        type: string
      envs:
        items:
          $ref: '#/definitions/http.DataSourceEnvResp'
        type: array
      id:
        type: integer
      is_dual:
        type: boolean
      name:
        type: string
      tenant_id:
        type: integer
      type:
        type: string
    type: object
  http.GatewayListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/http.GatewayResp'
        type: array
      pagination:
        $ref: '#/definitions/http.PaginationInfo'
    type: object
  http.GatewayResp:
    properties:
      id:
        type: integer
      ip:
        type: string
      last_seen:
        type: string
      name:
        type: string
      status:
        type: string
      tenant_id:
        type: integer
      token:
        type: string
      version:
        type: string
    type: object
  http.PaginationInfo:
    properties:
      is_pager:
        type: integer
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  http.R:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  http.RDataSource:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.DataSourceResp'
      msg:
        type: string
    type: object
  http.RDataSourceList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.DataSourceListResp'
      msg:
        type: string
    type: object
  http.RGateway:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.GatewayResp'
      msg:
        type: string
    type: object
  http.RGatewayList:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.GatewayListResp'
      msg:
        type: string
    type: object
...
  http.RTenant:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/http.TenantResp'
      msg:
        type: string
    type: object
...
  http.TenantResp:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: 企业级 SQL to API 智能网关平台管理接口
  title: OwlApi Control Plane
  version: "1.0"
paths:
...
  /v1/tenants/{slug}:
    delete:
      operationId: deleteTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除租户 (SuperAdmin)
      tags:
      - tenant
    get:
      operationId: getTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 获取租户详情
      tags:
      - tenant
    put:
      consumes:
      - application/json
      operationId: updateTenant
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 更新信息
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
            plan:
              type: string
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RTenant'
      security:
      - BearerAuth: []
      summary: 更新租户 (SuperAdmin)
      tags:
      - tenant
  /v1/tenants/{slug}/datasources:
    get:
      operationId: listDataSources
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RDataSourceList'
      security:
      - BearerAuth: []
      summary: 获取数据源列表
      tags:
      - datasource
    post:
      consumes:
      - application/json
      operationId: createDataSource
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 数据源信息
        in: body
        name: body
        required: true
        schema:
          properties:
            envs:
              type: array
            is_dual:
              type: boolean
            name:
              type: string
            type:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RDataSource'
      security:
      - BearerAuth: []
      summary: 创建数据源
      tags:
      - datasource
  /v1/tenants/{slug}/datasources/{datasourceId}:
    delete:
      operationId: deleteDataSource
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 数据源ID
        in: path
        name: datasourceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除数据源
      tags:
      - datasource
    get:
      operationId: getDataSource
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 数据源ID
        in: path
        name: datasourceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RDataSource'
      security:
      - BearerAuth: []
      summary: 获取数据源详情
      tags:
      - datasource
    put:
      consumes:
      - application/json
      operationId: updateDataSource
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 数据源ID
        in: path
        name: datasourceId
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: body
        schema:
          properties:
            envs:
              type: array
            is_dual:
              type: boolean
            name:
              type: string
            type:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RDataSource'
      security:
      - BearerAuth: []
      summary: 更新数据源
      tags:
      - datasource
  /v1/tenants/{slug}/gateways:
    get:
      operationId: listGateways
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: size
        type: integer
      - description: 是否分页，0=返回全部（默认1）
        in: query
        name: is_pager
        type: integer
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RGatewayList'
      security:
      - BearerAuth: []
      summary: 获取网关列表
      tags:
      - gateway
    post:
      consumes:
      - application/json
      operationId: createGateway
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 网关信息
        in: body
        name: body
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RGateway'
      security:
      - BearerAuth: []
      summary: 创建网关
      tags:
      - gateway
  /v1/tenants/{slug}/gateways/{gatewayId}:
    delete:
      operationId: deleteGateway
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 网关ID
        in: path
        name: gatewayId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.R'
      security:
      - BearerAuth: []
      summary: 删除网关
      tags:
      - gateway
    get:
      operationId: getGateway
      parameters:
      - description: 租户slug
        in: path
        name: slug
        required: true
        type: string
      - description: 网关ID
        in: path
        name: gatewayId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RGateway'
      security:
      - BearerAuth: []
      summary: 获取网关详情
      tags:
      - gateway
...
securityDefinitions:
  BearerAuth:
    description: '格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"